test: added validation logic tests into the corresponding script
